Split ColorMap Section out from layer card
